Ajout d'une verification de doublons lors de la creation de comptes
Ajout des liens d'actions (consulter, deposer, retirer, virer)

Creation et ajout du formulaire de creation de compte

+++ vue/include/compteAside.php
@@ -1,24 +1,30 @@
-<aside>
-    <div id="infos-compte-container">
-        <h2>Compte</h2>
-
-        <form action="" method="post">
-            <div class="form-group">
-                <select name="idCompte" id="">
-                    <option value="">--- Choisissez un compte ---</option>
-                </select>
-        </form>
-
-        <p><b>Solde :</b> -- €</p>
-        <p><b>Type de compte :</b> Courant</p>
-        <p><b>Client :</b> --</p>
-        <p><b>Identifiant :</b> --</p>
-    </div>
-
-    <!-- <div id="link-container">
+<?php if (isset($compte)): ?>
+    <aside>
+        <div id="infos-compte-container">
+            <h2>Compte</h2>
+            <form action="" method="post">
+                <div class="form-group">
+                    <select name="idCompte" id="">
+                        <option value="">--- Choisissez un compte ---</option>
+                        <?php foreach ($comptes as $compte): ?>
+                            <option value="<?= $compte->getId() ?>"><?= $compte->getId() ?></option>
+                        <?php endforeach; ?>
+                    </select>
+                </div>
+                <div class="form-group">
+                    <input type="submit" value="Selectionner">
+                </div>
+            </form>
+            <p><b>Solde :</b> <?= $compte->getSolde() ?></p>
+            <p><b>Type de compte :</b> <?= $compte->getType() ?></p>
+            <p><b>Client :</b> <?= $compte->getClient() ?></p>
+            <p><b>Identifiant :</b> <?= $compte->getId() ?></p>
+        </div>
+        <!-- <div id="link-container">
         <a href="?action=depot">Dépôt</a>
         <a href="?action=retrait">Retrait</a>
         <a href="?action=virement">Virement</a>
         <a href="?action=supprimerCompte">Supprimer le compte</a>
     </div> -->
-</aside>
+    </aside>
+<?php endif; ?>
